// cad-master/src/app/model/onefix-detail.ts
import { ProductComponent } from './product-component'
import { ProductVersion } from './product-version'
import { Product } from './product'

export class OneFixCase {
    // ID: number;
    CaseID: string;
    //Title: string;
    Description: string;
    // Resolution: string;
    Trace: string;
    // ProductDivision: string;
    // ProductLine: string;
    ServiceProduct: Product;
    ProductVersion: ProductVersion;
    Component: ProductComponent;
    // Status: string;
    // ResolutionType: string;
    // InitSupportCaseNumer: string;
    // EscEngineer: string;
    // DevEngineer: string;
    // Environment: string;
    // ReproSteps: string;
    // WorkAround: string;
    // UniqueSymptoms: string;
    // Severity: string;
    // LastUpdateTime: Date;
}